move MultipartUploadStream to its own file
